move  cpu specific logic to cpu_utils.py
