+++ python/sglang/srt/layers/moe/fused_moe_triton/layer.py
@@ -6,6 +6,7 @@ from typing import Callable, List, Optional, Tuple
 
 import torch
 
+from sglang.srt.cpu_utils import get_actual_shard_size
 from sglang.srt.custom_op import CustomOp
 from sglang.srt.distributed import (
     get_tensor_model_parallel_rank,
@@ -18,13 +19,7 @@ from sglang.srt.layers.quantization.base_config import (
     QuantizationConfig,
     QuantizeMethodBase,
 )
-from sglang.srt.utils import (
-    get_actual_shard_size,
-    get_bool_env_var,
-    is_hip,
-    permute_weight,
-    set_weight_attrs,
-)
+from sglang.srt.utils import get_bool_env_var, is_hip, permute_weight, set_weight_attrs
 
 if torch.cuda.is_available():
     from sglang.srt.layers.moe.fused_moe_triton.fused_moe import fused_experts
