+++ python/sglang/srt/utils.py
@@ -1451,7 +1451,3 @@ def set_cuda_arch():
         capability = torch.cuda.get_device_capability()
         arch = f"{capability[0]}.{capability[1]}"
         os.environ["TORCH_CUDA_ARCH_LIST"] = f"{arch}{'+PTX' if arch == '9.0' else ''}"
-
-
-def get_actual_shard_size(shard_size, weight_start, weight_end):
-    return min(shard_size, weight_end - weight_start)
